style: change arg order, implement for all cases
it now goes success/return callback, failure callback, and then *obj

Most times, the object won't be passed by the user, and can be handled
by the NexObject override.
In such a case, the user is more likely to add a failure callback than
a useless *obj that points to nothing.